+++ src/components/ChatForm.js
@@ -2,38 +2,20 @@ import React from 'react';
 import { FormInput } from './FormInput';
 import { DateMarker } from './DateMarker';
 import { MessageBox } from './MessageBox';
-import { TimeToDate } from './../lib/TimeToDate';
+import { ChatHeader } from './ChatHeader';
 import { ItIsThatDay } from './../lib/ItIsThatDay';
+import Parent from './Parent.Context';
 import styles from './../static/styles/ChatForm.module.css';
 
 export function ChatForm(props) {
+	const style = props.style;
 	const messageList = props.messageList;
 	const chatInfo = props.chatInfo;
 	const myInfo = props.myInfo;
 
-	let onlineStatus = 'групповой чат';
-	if (!chatInfo.isGroupChat) {
-		let online = chatInfo.penPal.status;
-		if (!online) {
-			onlineStatus = 'сейчас онлайн';
-		} else {
-			onlineStatus = `был в сети в ${TimeToDate(online)}`;
-		}
+	if (!chatInfo) {
+		return '';
 	}
-
-	let avatar = null;
-	try {
-		avatar = require(`../static/images/${chatInfo.avatar}`);
-	} catch {
-		avatar = require(`../static/images/default.png`);
-	}
-
-	let userImageStyles = {
-		backgroundImage: `url(${avatar})`,
-		backgroundRepeat: 'no-repeat',
-		backgroundPosition: 'center center',
-		backgroundSize: 'cover',
-	};
 
 	let lastTime = null;
 	let list = [];
@@ -48,35 +30,24 @@ export function ChatForm(props) {
 			Marker = <DateMarker key={$i++} time={currentTime} />;
 		}
 
-		if (currentTime >= lastTime) {
-			if (Marker) {
-				list.push(Marker);
-			}
-			list.push(Message);
-		} else {
-			list.unshift(Message);
-			if (Marker) {
-				list.unshift(Marker);
-			}
+		if (Marker) {
+			list.push(Marker);
 		}
+		list.push(Message);
 
 		lastTime = currentTime;
 	}
 
 	return (
-		<div className={styles.chatForm}>
-			<div className={styles.header}>
-				<div className={styles.headerButton + ' ' + styles.backButton}></div>
-				<div className={styles.nameConteiner}>
-					<div className={styles.userImage} style={userImageStyles}></div>
-					<div className={styles.userName}>
-						<div className={styles.name}>{chatInfo.chatName}</div>
-						<div className={styles.status}>{onlineStatus}</div>
-					</div>
-				</div>
-				<div className={styles.headerButton + ' ' + styles.searchButton}></div>
-				<div className={styles.headerButton + ' ' + styles.optionsButton}></div>
-			</div>
+		<div style={style} className={styles.chatForm}>
+			<Parent.Consumer>
+				{(value) => (
+					<ChatHeader
+						backToList={value.closeChat.bind(value)}
+						chatInfo={chatInfo}
+					/>
+				)}
+			</Parent.Consumer>
 			<div className={styles.content}>
 				<div className={styles.messageWrap}>
 					<DateMarker />
@@ -84,7 +55,14 @@ export function ChatForm(props) {
 				</div>
 			</div>
 			<div className={styles.footer}>
-				<FormInput placeholder="Ваше сообщение" />
+				<Parent.Consumer>
+					{(value) => (
+						<FormInput
+							formEntered={value.formEntered.bind(value)}
+							placeholder="Ваше сообщение"
+						/>
+					)}
+				</Parent.Consumer>
 			</div>
 		</div>
 	);
