+++ src/components/Parent.Context.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Parent = React.createContext(null);
+
+export default Parent;
