+++ src/static/styles/ChatHeader.module.css
@@ -1,0 +1,81 @@
+.header {
+	background-color: #2c2d2f;
+	width: 100%;
+	height: 60px;
+	flex-shrink: 0;
+	flex-grow: 0;
+	z-index: 1;
+	box-shadow: 0 0 2px 0 #151716;
+	display: flex;
+	flex-direction: row;
+}
+
+.headerButton {
+	height: 100%;
+	width: 30px;
+	margin: 0 15px;
+	cursor: pointer;
+}
+
+.nameConteiner {
+	flex: auto;
+	height: 100%;
+	padding: 8px 0;
+	display: flex;
+	flex-direction: row;
+	justify-content: center;
+}
+
+.userImage {
+	height: 44px;
+	width: 44px;
+	border-radius: 25px;
+	margin-right: 15px;
+}
+
+.userName {
+	/*flex: auto;*/
+	height: 100%;
+	padding-top: 5px;
+	text-shadow: 1px 1px 0 #222;
+}
+
+.name {
+	font-size: var(--fontMinSize);
+}
+
+.status {
+	font-size: var(--fontMinMinSize);
+	color: #9a9b9d;
+}
+
+.backButton {
+	background: url(../images/back.png) no-repeat center center;
+	background-size: 100%;
+}
+
+.searchButton {
+	background: url(../images/search.png) no-repeat center center;
+	background-size: 90%;
+	margin-right: 0;
+}
+
+.optionsButton {
+	background: url(../images/options.png) no-repeat center center;
+	background-size: 90%;
+}
+
+.headerButton {
+	height: 100%;
+	width: 30px;
+	opacity: 0.5;
+	transition-duration: 0.15s;
+}
+
+.headerButton:hover {
+	opacity: 1;
+}
+
+.headerButton:active {
+	opacity: 0.7;
+}
