+++ src/static/styles/Main.module.css
@@ -5,22 +5,6 @@
 	background: url(../images/backgroundMain.png);
 }
 
-.chat {
-	z-index: 1;
-	right: -100%;
-	animation-duration: 0.5s;
-	animation-timing-function: ease-in-out;
-	animation-fill-mode: forwards;
-}
-
-.apear {
-	animation-name: chatApearence;
-}
-
-.disapear {
-	animation-name: chatDisapear;
-}
-
 @keyframes chatDisapear {
 	from {
 		right: 0;
@@ -33,7 +17,7 @@
 	}
 }
 
-@keyframes chatApearence {
+@keyframes chatApear {
 	from {
 		right: -100%;
 		/*opacity: 0.9;*/
