+++ src/static/styles/MessageBox.module.css
@@ -1,7 +1,7 @@
 .messageConteiner {
 	width: 100%;
 	box-sizing: border-box;
-	padding: 0 10px 20px 10px;
+	padding: 0 10px 10px 10px;
 	width: 100%;
 	display: block;
 	overflow: hidden;
